refactor: 핵심 로직을 가진 RacingGame 객체를 domain 패키지로 이동
