+++ src/test/java/service/RacingGameTest.java
@@ -4,6 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import domain.Car;
+import domain.RacingGame;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.LinkedList;
